requirements.txt
Django==3.2.8
djangorestframework==3.12.3
markdown
django-filter
sphinx
whitenoise
paho-mqtt
django-background-tasks
